Quell various compiler warnings.

+++ smtk/bridge/rgg/plugin/smtkRGGAddMaterialView.cxx
@@ -249,7 +249,7 @@ bool smtkRGGAddMaterialView::ableToOperate()
     {
       return false;
     }
-    densityTypeI->setDiscreteIndex(index);
+    densityTypeI->setDiscreteIndex(static_cast<int>(index));
   }
 
   smtk::attribute::StringItemPtr compositionTypeI =
@@ -265,7 +265,7 @@ bool smtkRGGAddMaterialView::ableToOperate()
     {
       return false;
     }
-    compositionTypeI->setDiscreteIndex(index);
+    compositionTypeI->setDiscreteIndex(static_cast<int>(index));
   }
 
   smtk::attribute::StringItemPtr componentI =
